Fixed: token volume series overflow

+++ internal/models/transfer/repository.go
@@ -8,5 +8,5 @@ type Repository interface {
 	GetAll(network string, level int64) ([]Transfer, error)
 	GetTokenSupply(network, address string, tokenID int64) (result TokenSupply, err error)
 	GetToken24HoursVolume(network, contract string, initiators, entrypoints []string, tokenID int64) (float64, error)
-	GetTokenVolumeSeries(network, period string, contracts []string, entrypoints []tzip.DAppContract, tokenID uint) ([][]int64, error)
+	GetTokenVolumeSeries(network, period string, contracts []string, entrypoints []tzip.DAppContract, tokenID uint) ([][]float64, error)
 }
